cloud_run_service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
google-genai>=1.0
pydantic>=2.5
